updated sample data with my user ID for testing. Moved files.
