+++ src/components/Authenticated.js
