+++ src/components/SignIn.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { signInUser } from '../api/auth';
+import { signInUser, signOutUser } from '../api/auth';
 
 export default function SignIn() {
   return (
@@ -7,6 +7,9 @@ export default function SignIn() {
       <h1>Welcome! Sign In!</h1>
       <button type="button" className="btn btn-success" onClick={signInUser}>
         Sign In
+      </button>
+      <button type="button" className="btn btn-danger" onClick={signOutUser}>
+        Sign Out
       </button>
     </div>
   );
